PoC: ansible_module.turbo integration
This PoC PR introduces the use of ansible_module.turbo[1] to speed up the
`k8s_info` module. It breaks everything else :-).

This two recording runs the `k8s_facts` 20 times in row. The first one
use the current `main` branch, the second use this PR.

- https://asciinema.org/a/364880: 21.2s
- https://asciinema.org/a/364881: 7.2s

This PR improves performance at two level.

Reuse the same client object
============================

With ansible_module.turbo, the module are actually run in a long running
Python process. Thanks to it, we can cache large Python objects. In this case,
we focus on the `openshift.dynamic.DynamicClient` instance.
We use the `lru_cache`[2] decorator to ensure we will always get the
same python object, as soon as we use the same authentication
parameters.
Everytime we call `get_api_client()` in a new task, we get the same
Python client that is already in memory.

Speed up the import
===================

The loading of the OpenShift SDK take a lot of time, and this is a
penality we pay everytime we start a new task:

```
$ time python3 -c "from openshift.dynamic import DynamicClient"

real    0m0.653s
user    0m0.966s
sys     0m0.469s
```

This PR try to defeer all the non-critical imports. This allow us to load
the modules once in the long running Python process. We can reuse the
same module task after task.

About this PR changes
=====================

The use of object complexify the ansible_module.turbo integration.
Currently, everytime we instanciate the `K8sAnsibleMixin` object we load all the SDK.
In addition, everytime we run a task, we end up with a new
`K8sAnsibleMixin` instance that can hardly cache.

These are the reasons why the flatten the nested structure.

If you like the approach, I will be happy to start a larger refactoring
to generalize the use of ansible_module.turbo.

- [1] https://github.com/ansible-collections/cloud.common/blob/main/README_ansible_turbo.module.rst
- [2] https://docs.python.org/3/library/functools.html#functools.lru_cache

Signed-off-by: Gonéri Le Bouder <goneri@lebouder.net>